Refactor: improve SauceDemo tests and extensions for clarity and maintainability
This commit refactors the SauceDemo tests and extension methods to improve clarity, maintainability, and organization by using partial classes and descriptive summaries.

- Refactored `SauceDemoTests` into partial classes (`Base`, `Data`, `Login`, `Inventory`) to improve organization and maintainability.
  - Moved test data to `SauceDemoTests.Data.cs`.
  - Created `SauceDemoTests.Login.cs` to focus on login scenarios, including the "problem_user" test.
  - Created `SauceDemoTests.Inventory.cs` to focus on inventory page interactions.
  - Updated `SauceDemoTests.Base.cs` to include core fixture setup and SauceDemo-specific initializations.
- Updated `ShouldDisplaySameImageForAllProductsForProblemUser` test to verify the "problem_user" scenario, where all product images are the same.
- Refactored `ExtensionMethods` into partial classes (`Base`, `Dropdown`, `Login`) for better organization.
- Added comprehensive summaries and remarks to the `SauceDemoTests` and `ExtensionMethods` classes to improve documentation and understanding.
- Updated `ExtensionMethods.Base.cs` with a detailed description of the extension methods and their organization.
- Removed obsolete summaries from `ExtensionMethods.Dropdown.cs` and `ExtensionMethods.Login.cs`.
